adaptation to IEMOCAP

+++ src/dataset.py
@@ -54,6 +54,6 @@ class Multimodal_Datasets(Dataset):
         if self.data == 'mosi':
             META = (self.meta[index][0].decode('UTF-8'), self.meta[index][1].decode('UTF-8'), self.meta[index][2].decode('UTF-8'))
         if self.data == 'iemocap':
-            Y = torch.argmax(Y, dim=-1)
+            # Convert per-emotion logits to binary targets of shape (4,)
+            Y = torch.argmax(Y, dim=-1).long()
         return X, Y, META        
-
